[Config Save] Error happen when config save a filename with quotation mark
When the filename passed to the CLI command 'config save' contains the
special character - double quote, it needs to add an escape character
'\' before the double quote to let bash treat it as a literal character.

However, the filename passed into the python code also contains the
escape character and it will cause a 'file-not-found' error when simply
passes that filename to 'read_json_file()'.

To fix the error, use 'echo' command with the filename to strip the
special character.